more cleanup of match logic
